sgaq/src/requirements.txt
mysql-connector-python
tk
cryptography
